+++ package.json
@@ -36,6 +36,7 @@
     "lerna": "^4.0.0",
     "lint-staged": "^10.5.4",
     "prettier": "^2.2.1",
+    "rimraf": "^3.0.2",
     "typescript": "^4.2.4"
   },
   "browserslist": {
