+++ packages/ui/package.json
@@ -42,7 +42,9 @@
     "body-scroll-lock": "^3.1.5",
     "classnames": "^2.2.6",
     "react-input-mask": "^2.0.4",
-    "react-transition-group": "^4.4.1"
+    "react-transition-group": "^4.4.1",
+    "@eruditorgroup/profi-icons": "0.0.0",
+    "@eruditorgroup/profi-toolkit": "0.0.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.13.10",
